Refactor findFilepath
