+++ src/util.js
@@ -136,6 +136,35 @@ const signatureToCompletion = (signatures, kind, detail) => {
   return completionSet;
 };
 
+/**
+ * Checks a filename with the include paths for a valid path
+ * @param {string} file - the filename to append to the paths
+ * @param {boolean} library - Search Autoit library first?
+ * @returns {(string|boolean)} Full path if found to exist or false
+ */
+ const findFilepath = (file, library = true) => {
+  // work with copy to avoid changing main config
+  const includePaths = [...workspace.getConfiguration('autoit').get('includePaths')];
+  if (!library) {
+	// move main library entry to the bottom so that it is searched last
+    includePaths.push(includePaths.shift());
+  } 
+
+  let newPath;
+  const pathFound = includePaths.some(iPath => {
+    newPath = path.normalize(`${iPath}\\`) + file;
+    if (fs.existsSync(newPath)) {
+      return true;
+    }
+    return false;
+  });
+
+  if (pathFound && newPath) {
+    return newPath;
+  }
+  return false;
+};
+
 module.exports = {
   descriptionHeader,
   valueFirstHeader,
@@ -155,4 +184,5 @@ module.exports = {
   functionPattern,
   completionToHover,
   signatureToCompletion,
+  findFilepath,
 };
